large changes to overal design plus working breathe function
For those reading, the breathe function requires the % operator on floats
in theory I could have re-written it without floats though that would be a pain
either I use fixed point numbers (very possibly would work) or do something even more cursed
I'm going to do a bit more research into this issue with floats since it is a pretty big one
compile some code for the raspberry pi 1 b (arm-unknown-linux-gnueabi) to test if
float modulo is really the problem. Anyway enough of this, other changes were
pretty good, I'm starting to doc string a bunch of functions, I know a bunch won't be public
but it's good practice for in the future especially with writing low level code.
lots of moving functionality to it's own function this should mean certain presets are really
easily done.
Got some more cleaning up to do and starting to parralelize / async it so I can run a
server to take input instead of command line input :/